+++ NEWS.md
@@ -14,6 +14,7 @@
 * `mrClosed()`: Modified. Added some checks for non-vector uses of `M=` and `R=` (partially addresses [#22](https://github.com/droglenc/FSA/issues/22)).
 * `predict.bootCase()`: Modified. Modified so that situations where other than values of the dependent variable are in the dots argument (as would occur if making predictions for the Francis parameterization of the VBGF).
 * `predict.nlsBoot()`: Modified. See note for `predict.bootCase()`.
+* `psdCalc()`: Modified. Fixed bug in output if more than two additional lengths were supplied.
 * `removal()`: Modified. Fixed bug related to sending catches in a one column data.frame. Fixed bug related to selecting only one `parm=` in `confint()`.  Added tests.
 * `residPlot()`: Modified.  Added `cex.leg=` and `box.lty.leg=` to IVR plots.
 * `vbFuns()`: Modified.  Added `Ogle` to list of parameterizations. Changed order of `L0` and `K` parameters in returned function when `param="Original"`.